Start this process over
This method did not work because I was trying to increment the innerHTML of the elements as if were a number, but the fact that it was written in a paragraph tag made it a string in JavaScript.

Will try again, this time leaving the innerHTML blank in index.html and filling it in with JavaScript so I can set it as a number.