Set computer choice variable
Created the function compChoose and named the var compChoice inside of it.

Used Math.random to generate a random number in between 0 and 1.

Multiplied it by 3 to get a number in between 0 and 2.9999...

Used Math.ceiling to round up to the nearest integer so it returns either 1, 2, or 3.

Used an else if loop to set the var compChoice to rock if number is 1, paper if number is 2, and scissors if number is 3.