Transitions
Set .choice to scale(1.25) when hovered over.

Set its transition time to .1s sec but the transition time of .choice:hover to .5 sec so it will ease in and snap out.

Used an else if statement to add the class .scale, which applies scale(1.25), to #questMark when the function game is ran.

Removed .scale from #questMark before adding it in the else if loop so it will revert back to its normal size first, giving it the ease-in effect.