Fixed reset() function
The previous edit fixed the problem of the score counters innerHTML being undefined after being loaded into a new browser, but it also resulted in them resetting to 0 everytime the page was refreshed.

To fix this, I changed the onload reset() function to an onload load() function that sets the innerHTML to 0 if the local storage is undefined.